feat: qa_base class
